Split out Permission from authorization::mod.rs
This commit moves the `Permission` enum into its own file out of the
mod.rs file. The goal is to keep mod.rs concise and more tightly
focused.
